+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.2.16
+
+- `async_extension`:
+  - Added export of `src/async_extension_compute_once.dart` to library.
+  - New `ComputeOnce`.
+
+- Dependencies:
+  - Updated `test` dependency from ^1.26.0 to ^1.26.3.
+  - Updated `coverage` dependency from ^1.13.1 to ^1.15.0.
+
 ## 1.2.15
 
 - New `FunctionArgs0Extension`, `FunctionArgs1Extension`, `FunctionArgs2Extension`: